Fix: correct gc_depth type mismatch and restore missing definitions

+++ code/bullshark/consensus/core/src/speculative_attack_test.rs
@@ -129,7 +129,7 @@ async fn test_speculative_attack_asr_dynamic() {
         .ok()
         .and_then(|s| s.parse().ok())
         .unwrap_or(10);
-    protocol_config.set_consensus_gc_depth_for_testing(gc_depth);
+    protocol_config.set_consensus_gc_depth_for_testing(gc_depth as u32);
 
     let temp_dirs = (0..num_validators)
         .map(|_| TempDir::new().unwrap())
